Add usage example
Restore package changes

Cleanup PrintResult codes
