+++ client_test.go
@@ -4,13 +4,13 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-package nebula
+package ngdb
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/vesoft-inc/nebula-go/nebula/graph"
+	"github.com/vesoft-inc/nebula-go/graph"
 )
 
 const (
@@ -20,6 +20,11 @@ const (
 	password = "password"
 )
 
+// Before run `go test -v`, you should start a nebula server listening on 6699 port.
+// Using docker-compose is the easiest way and you can reference this file:
+//   https://github.com/vesoft-inc/nebula/blob/master/docker/docker-compose.yaml
+//
+// TODO(yee): introduce mock server
 func TestClient(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skipping client test in short mode")
@@ -39,16 +44,14 @@ func TestClient(t *testing.T) {
 	if resp, err := client.Execute("SHOW HOSTS;"); err != nil {
 		t.Errorf(err.Error())
 	} else {
-		result := client.PrintResult(resp)
-		t.Logf("show hosts:\n%s\nErrorCode: %v, ErrorMsg: %s", result, resp.GetErrorCode(), resp.GetErrorMsg())
+		t.Logf("show hosts, ErrorCode: %v, ErrorMsg: %s", resp.GetErrorCode(), resp.GetErrorMsg())
 	}
 
 	if resp, err := client.Execute("CREATE SPACE client_test(partition_num=1024, replica_factor=1);"); err != nil {
 		t.Error(err.Error())
 	} else {
-		t.Logf("create space response: %s", client.PrintResult(resp))
 		if resp.GetErrorCode() != graph.ErrorCode_SUCCEEDED {
-			t.Logf("create space: ErrorCode: %v, ErrorMsg: %s", resp.GetErrorCode(), resp.GetErrorMsg())
+			t.Logf("create space, ErrorCode: %v, ErrorMsg: %s", resp.GetErrorCode(), resp.GetErrorMsg())
 		}
 	}
 }
