+++ graph/constants.go
