+++ graph/ttypes.go
